+++ CHANGELOG.md
@@ -98,6 +98,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   can be customized using `format_detailed_cryptographic_signature(signature)` and
   `format_short_cryptographic_signature(signature)`.
 
+* New `git.init-track-local-bookmarks` config option to automatically set local
+  bookmarks as tracking remote bookmarks during `jj git init --colocate`.
+
 ### Fixed bugs
 
 * Fixed diff selection by external tools with `jj split`/`commit -i FILESETS`.